Return books instead of strings from search
This commit causes the library's search feature to return an array of
book objects instead of just the titles. Anywhere that was expecting an
array of strings has been updated to extract that string, and from now
on this affords more flexibility in displaying the data since everything
is there to choose from.